requirements.txt
datasets
pandas
Pillow
tensorboard
torch
torchmetrics
torchsummary
torchvision